+++ aerospike-crud-store/src/main/java/com/livetheoogway/crudstore/aerospike/AerospikeStore.java
@@ -21,40 +21,53 @@ import com.aerospike.client.Key;
 import com.aerospike.client.Record;
 import com.aerospike.client.policy.RecordExistsAction;
 import com.aerospike.client.policy.WritePolicy;
+import com.aerospike.client.query.Filter;
+import com.aerospike.client.query.IndexCollectionType;
+import com.aerospike.client.query.IndexType;
+import com.aerospike.client.query.RecordSet;
+import com.aerospike.client.query.Statement;
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.livetheoogway.crudstore.core.Id;
-import com.livetheoogway.crudstore.core.Store;
+import com.livetheoogway.crudstore.core.ReferenceExtendedStore;
 import lombok.extern.slf4j.Slf4j;
-import lombok.val;
 
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 @Slf4j
-public abstract class AerospikeStore<T extends Id> implements Store<T> {
+public abstract class AerospikeStore<T extends Id> implements ReferenceExtendedStore<T> {
 
-    protected static final String DATA = "data";
+    private static final String DEFAULT_DATA_BIN = "data";
+    private static final String DEFAULT_REF_ID_BIN = "refId";
+    private static final String DEFAULT_REF_ID_INDEX_SUFFIX = "_idx";
+
     protected final ObjectMapper mapper;
     protected final IAerospikeClient client;
     protected final NamespaceSet namespaceSet;
     protected final WritePolicy createPolicy;
     protected final WritePolicy updateOnly;
     protected final ErrorHandler<T> errorHandler;
-    private final Class<T> clazz;
+    protected final TypeReference<T> typeReference;
+    protected final AerospikeStoreSetting storeSetting;
 
     protected AerospikeStore(final IAerospikeClient client,
                              final NamespaceSet namespaceSet,
                              final ObjectMapper mapper,
                              final Class<T> clazz,
                              final ErrorHandler<T> errorHandler) {
-        this(client, namespaceSet, mapper, clazz, errorHandler, true);
+        this(client, namespaceSet, mapper, clazz, errorHandler, null);
     }
 
     protected AerospikeStore(final IAerospikeClient client,
@@ -62,17 +75,42 @@ public abstract class AerospikeStore<T extends Id> implements Store<T> {
                              final ObjectMapper mapper,
                              final Class<T> clazz,
                              final ErrorHandler<T> errorHandler,
-                             final boolean failOnCreateIfRecordExists) {
+                             final AerospikeStoreSetting storeSetting) {
+        this(client, namespaceSet, mapper, new TypeReference<>() {
+            @Override
+            public Type getType() {
+                return clazz;
+            }
+        }, errorHandler, storeSetting);
+    }
+
+    protected AerospikeStore(final IAerospikeClient client,
+                             final NamespaceSet namespaceSet,
+                             final ObjectMapper mapper,
+                             final TypeReference<T> typeReference,
+                             final ErrorHandler<T> errorHandler) {
+        this(client, namespaceSet, mapper, typeReference, errorHandler, null);
+    }
+
+    protected AerospikeStore(final IAerospikeClient client,
+                             final NamespaceSet namespaceSet,
+                             final ObjectMapper mapper,
+                             final TypeReference<T> typeReference,
+                             final ErrorHandler<T> errorHandler,
+                             final AerospikeStoreSetting storeSetting) {
         this.client = client;
         this.namespaceSet = namespaceSet;
         this.mapper = mapper;
         this.errorHandler = errorHandler;
-        this.clazz = clazz;
+        this.typeReference = typeReference;
         this.createPolicy = new WritePolicy(client.getWritePolicyDefault());
-        createPolicy.recordExistsAction = failOnCreateIfRecordExists ? RecordExistsAction.CREATE_ONLY
-                                                                     : RecordExistsAction.REPLACE;
+        this.storeSetting = defaultIfNull(storeSetting, namespaceSet.set());
+        createPolicy.recordExistsAction = this.storeSetting.failOnCreateIfRecordExists()
+                                          ? RecordExistsAction.CREATE_ONLY
+                                          : RecordExistsAction.REPLACE;
         this.updateOnly = new WritePolicy(client.getWritePolicyDefault());
         updateOnly.recordExistsAction = RecordExistsAction.UPDATE_ONLY;
+        setupIndexes();
     }
 
     /**
@@ -102,7 +140,12 @@ public abstract class AerospikeStore<T extends Id> implements Store<T> {
      */
     @Override
     public void create(final T item) {
-        write(item, createPolicy);
+        write(item.id(), () -> recordDetails(item, null), createPolicy);
+    }
+
+    @Override
+    public void create(final T item, List<String> refIds) {
+        write(item.id(), () -> recordDetails(item, refIds), createPolicy);
     }
 
     /**
@@ -112,14 +155,14 @@ public abstract class AerospikeStore<T extends Id> implements Store<T> {
      */
     @Override
     public void update(final T item) {
-        write(item, updateOnly);
+        write(item.id(), () -> recordDetails(item, null), updateOnly);
     }
 
 
     @Override
     public void delete(final String id) {
         exec("delete", id, () -> {
-            final Key requestIdKey = new Key(namespaceSet.namespace(), namespaceSet.set(), id);
+            final var requestIdKey = new Key(namespaceSet.namespace(), namespaceSet.set(), id);
             if (!client.delete(updateOnly, requestIdKey)) {
                 errorHandler.onDeleteUnsuccessful();
             }
@@ -135,15 +178,11 @@ public abstract class AerospikeStore<T extends Id> implements Store<T> {
      */
     @Override
     public Optional<T> get(final String id) {
-        final T data = exec("get", id, () -> {
-            final Key requestIdKey = new Key(namespaceSet.namespace(), namespaceSet.set(), id);
-            final Record asRecord = client.get(client.getReadPolicyDefault(), requestIdKey);
-            return extractItem(id, asRecord);
+        return exec("get", id, () -> {
+            final var requestIdKey = new Key(namespaceSet.namespace(), namespaceSet.set(), id);
+            final var asRecord = client.get(client.getReadPolicyDefault(), requestIdKey);
+            return extractItemFromRecord(id, asRecord);
         }, errorHandler);
-        if (isValidDataItem(data)) {
-            return Optional.ofNullable(data);
-        }
-        return Optional.empty();
     }
 
     /**
@@ -156,13 +195,13 @@ public abstract class AerospikeStore<T extends Id> implements Store<T> {
      */
     @Override
     public Map<String, T> get(final List<String> ids) {
-        final Key[] batchReads = ids.stream()
+        final var batchReads = ids.stream()
                 .map(id -> (new Key(namespaceSet.namespace(), namespaceSet.set(), id)))
                 .toArray(Key[]::new);
         final Record[] records = client.get(client.getBatchPolicyDefault(), batchReads);
         return IntStream
                 .range(0, ids.size())
-                .mapToObj(index -> extractItemForBulkOperations(batchReads[index].userKey.toString(), records[index]))
+                .mapToObj(index -> extractItemFromRecord(batchReads[index].userKey.toString(), records[index]))
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .filter(this::isValidDataItem)
@@ -183,39 +222,62 @@ public abstract class AerospikeStore<T extends Id> implements Store<T> {
                 client.getScanPolicyDefault(),
                 namespaceSet.namespace(), namespaceSet.set(),
                 (key, asRecord) ->
-                        extractItemForBulkOperations(key.userKey.toString(), asRecord)
+                        extractItemFromRecord(key.userKey.toString(), asRecord)
                                 .ifPresent(items::add));
         return items;
     }
 
-    public T extractItem(String id, Record asRecord) {
+    @Override
+    public List<T> getByRefId(final String refId) {
+        final var statement = new Statement();
+        statement.setNamespace(namespaceSet.namespace());
+        statement.setSetName(namespaceSet.set());
+        statement.setIndexName(storeSetting.refIdIndex());
+        statement.setBinNames(storeSetting.dataBin());
+        statement.setFilter(Filter.contains(storeSetting.refIdBin(), IndexCollectionType.LIST, refId));
+        final List<T> results = new ArrayList<>();
+        try (final RecordSet recordSet = client.query(client.getQueryPolicyDefault(), statement)) {
+            while (recordSet.next()) {
+                extractItemFromRecord(recordSet.getKey().userKey.toString(), recordSet.getRecord())
+                        .ifPresent(results::add);
+            }
+        } catch (AerospikeException e) {
+            log.error("[{}] Aerospike Error {} item for id:{}", typeReference.getType().getTypeName(), "getByRefId", refId, e);
+            return errorHandler.onAerospikeErrorForRefId(refId, e);
+        }
+        return results;
+    }
+
+    /**
+     * given an Aerospike Record, extract the data out of it
+     * @param id       key
+     * @param asRecord how the record is extracted
+     * @return optionally return data if available
+     */
+    protected Optional<T> extractItemFromRecord(final String id, final Record asRecord) {
         if (asRecord == null) {
             return errorHandler.onNoRecordFound(id);
         }
-        val data = asRecord.getString(DATA);
+        final var data = asRecord.getString(storeSetting.dataBin());
         try {
-            return mapper.readValue(data, clazz);
+            final T value = mapper.readValue(data, typeReference);
+            if (isValidDataItem(value)) {
+                return Optional.of(value);
+            }
+            log.warn("Invalid item found for id:{}", id);
+            return Optional.empty();
         } catch (JsonProcessingException e) {
-            log.error("[{}] Deserialization error id:{} record:{}", clazz.getSimpleName(), id, asRecord, e);
+            log.error("[{}] Deserialization error id:{} record:{}", typeReference.getType().getTypeName(), id, asRecord, e);
             return errorHandler.onDeSerializationError(id, e);
         }
     }
 
-    public Optional<T> extractItemForBulkOperations(String id, Record asRecord) {
-        if (asRecord == null) {
-            return errorHandler.onNoRecordFoundForBulkGet(id);
+    protected RecordDetails recordDetails(final T item, final List<String> refIds) throws JsonProcessingException {
+        final var dataBin = new Bin(storeSetting.dataBin(), mapper.writeValueAsString(item));
+        if (refIds != null && !refIds.isEmpty()) {
+            final var refIdBin = new Bin(storeSetting.refIdBin(), refIds);
+            return new RecordDetails(expiration(item), dataBin, refIdBin);
         }
-        val data = asRecord.getString(DATA);
-        try {
-            return Optional.of(mapper.readValue(data, clazz));
-        } catch (JsonProcessingException e) {
-            log.error("[{}] Deserialization error id:{} record:{}", clazz.getSimpleName(), id, asRecord, e);
-            return errorHandler.onDeSerializationErrorDuringBulkGet(id, e);
-        }
-    }
-
-    protected RecordDetails recordDetails(T item) throws JsonProcessingException {
-        final Bin dataBin = new Bin(DATA, mapper.writeValueAsString(item));
         return new RecordDetails(expiration(item), dataBin);
     }
 
@@ -228,12 +290,15 @@ public abstract class AerospikeStore<T extends Id> implements Store<T> {
      * @param <R>       result type
      * @return result
      */
-    protected <R> R exec(final String operation, final String id, ESupplier<R> response, ErrorHandler<R> errorHandler) {
+    protected <R> Optional<R> exec(final String operation,
+                                   final String id,
+                                   final ESupplier<Optional<R>> response,
+                                   final ErrorHandler<R> errorHandler) {
         try {
-            log.info("[{}] {} item for id:{}", clazz.getSimpleName(), operation, id);
+            log.info("[{}] {} item for id:{}", typeReference.getType().getTypeName(), operation, id);
             return response.get();
         } catch (AerospikeException e) {
-            log.error("[{}] Aerospike Error {} item for id:{}", clazz.getSimpleName(), operation, id, e);
+            log.error("[{}] Aerospike Error {} item for id:{}", typeReference.getType().getTypeName(), operation, id, e);
             return errorHandler.onAerospikeError(id, e);
         } catch (JsonProcessingException e) {
             log.error("Error while converting to string for id:{}", id, e);
@@ -244,22 +309,58 @@ public abstract class AerospikeStore<T extends Id> implements Store<T> {
         }
     }
 
-    private void write(final T item, WritePolicy defaultWritePolicy) {
-        val id = item.id();
+    protected void write(final String id,
+                         final ESupplier<RecordDetails> recordDetailsSupplier,
+                         final WritePolicy defaultWritePolicy) {
         exec("operation:" + defaultWritePolicy.recordExistsAction, id, () -> {
-            val recordDetails = recordDetails(item);
+            final var recordDetails = recordDetailsSupplier.get();
             var writePolicy = defaultWritePolicy;
             if (recordDetails.expiration() > 0) {
                 writePolicy = new WritePolicy(writePolicy);
                 writePolicy.expiration = recordDetails.expiration();
             }
-            val requestIdKey = new Key(namespaceSet.namespace(), namespaceSet.set(), id);
+            final var requestIdKey = new Key(namespaceSet.namespace(), namespaceSet.set(), id);
             client.put(writePolicy, requestIdKey, recordDetails.bins());
             return null;
         }, errorHandler);
     }
 
-    private interface ESupplier<T> {
-        T get() throws Exception;
+    private void setupIndexes() {
+        try {
+            final var refIdIndexTask = client
+                    .createIndex(null, namespaceSet.namespace(), namespaceSet.set(),
+                                 storeSetting.refIdIndex(), storeSetting.refIdBin(),
+                                 IndexType.STRING,
+                                 IndexCollectionType.LIST);
+            if (refIdIndexTask != null) {
+                refIdIndexTask.waitTillComplete();
+                log.info("Created index: {}", storeSetting.refIdIndex());
+            }
+        } catch (AerospikeException e) {
+            if (e.getResultCode() == 100 || e.getResultCode() == 200) {
+                log.info("Index already exists:{}", storeSetting.refIdIndex());
+                return;
+            }
+            log.error("Error while creating index:{}", storeSetting.refIdIndex(), e);
+            throw e;
+        }
+    }
+
+    private AerospikeStoreSetting defaultIfNull(final AerospikeStoreSetting storeSetting,
+                                                final @NotNull @NotEmpty String set) {
+        final var builder = AerospikeStoreSetting.builder();
+        if (storeSetting == null) {
+            return builder
+                    .dataBin(DEFAULT_DATA_BIN)
+                    .refIdBin(DEFAULT_REF_ID_BIN)
+                    .failOnCreateIfRecordExists(true)
+                    .refIdIndex(set + DEFAULT_REF_ID_INDEX_SUFFIX).build();
+        }
+        builder.dataBin(Objects.requireNonNullElse(storeSetting.dataBin(), DEFAULT_DATA_BIN));
+        final var refIdBin = Objects.requireNonNullElse(storeSetting.refIdBin(), DEFAULT_REF_ID_BIN);
+        builder.refIdBin(refIdBin);
+        builder.refIdIndex(Objects.requireNonNullElse(storeSetting.refIdIndex(),
+                                                      set + "_" + refIdBin + DEFAULT_REF_ID_INDEX_SUFFIX));
+        return builder.build();
     }
 }